requirements.txt
Django==4.0.5
djangorestframework==3.13.1
djongo==1.3.6
requests==2.28.1
uritemplate==4.1.1
xhtml2pdf==0.2.8
django-cors-headers==3.13.0
pymongo[srv]==3.12.3
django-heroku==0.3.1
gunicorn==20.0.4
whitenoise==5.2.0
python-dotenv==0.15.0
djangorestframework-simplejwt